fix: look up the recipient DEK, not your own DEK (#4632)
### Description

previously, anyone who sent with a comment after verifying their own
phone number with CPV would erroneously encrypt the comment with their
own DEK, rather than that of the recipient. The comment appeared as
"Comment Unavailable" for the recipient, who failed to decode it.

### Test plan

- fixed unit test that will now provide regression coverage
- manually tested. sent a transaction from a wallet with a verified
phone number and confirmed it can be decoded now by another wallet (and
that it is still encoded on chain). tx is here
https://celoscan.io/tx/0xe766b83c0798880aa233b74da1a0c1a3c9c7296f97119d869d6ee15bba29635d
.

### Related issues

na

### Backwards compatibility

na