refactor(swap): use useReducer to make SwapScreen state changes easier to reason about (#4673)
### Description

Switching to `useReducer` instead of many individual `useState` to
clarify possible/allowed state changes.

A couple of notes:
- This should make it harder to do invalid state changes. And facilitate
further changes.
- They all need to be triggered via defined actions (modelled as events)
- I was able to remove the state for `shouldShowMaxSwapAmountWarning`
and instead derive it from the value entered.
- The only difference is the warning will now show before hitting
"Confirm Swap". But I think this is good.
- `fromToken` and `toToken` are now always up-to-date (for instance in
case of balance changes), as we now only store the token IDs in the
state
- We could argue the reducer could be in the global redux reducer for
swap, but I think this is really component state. For instance, it
doesn't really make sense to trigger a global redux action every time a
character is entered.
- Also, with further changes, we could even stop going through redux for
executing the swap.

See also: [Comparing useState and
useReducer](https://react.dev/learn/extracting-state-logic-into-a-reducer#comparing-usestate-and-usereducer)

Next up: making it easier to understand which warning is displayed.

### Test plan

- Tests pass
- Manual tests

### Related issues

- Fixes RET-950

### Backwards compatibility

Yes