fix(onboarding): revert default of CPV screen to be in onboarding (#4876)
### Description

Fixes a bug introduced in #4834 , where VerificationStartScreen
defaulted to being not in onboarding. This causes an issue if a user
bails on the screen during onboarding. On restart, the screen is loaded
in the already onboarded mode where the back button does nothing and
there's no option to skip (see before video below). This PR reverts the
default to be in onboarding as it doesn't seem trivial to pass props on
the initialRoute.

### Test plan

Unit tests and manual

Before:


https://github.com/valora-inc/wallet/assets/5062591/496876ec-4adf-48b8-8a12-72d9030f3e37


After:


https://github.com/valora-inc/wallet/assets/5062591/8d345432-4fcb-439f-b63a-03806d5959e0


### Related issues

N/A

### Backwards compatibility

Yes