feat: add jumpstart button to select recipients screen (#4965)
### Description

This PR adds the remote config and button for the start of the escrow
jumpstart flow. I think that the most straightforward way to implement
the escrow flow is to use the existing send flow rather than separate
the flows. I've made a POC that is working e2e in #4964, in case it
helps to understand this decision.

### Test plan

![Simulator Screenshot - iPhone 14 Pro - 2024-02-23 at 12 22
01](https://github.com/valora-inc/wallet/assets/20150449/8b524157-4c52-481b-af68-b7a3d47885ca)


### Related issues

- Related to RET-993

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)