fix(send): show correct address in confirmation screen (#5019)
### Description

Fixes an edge case where the send confirmation address displays the
incorrect address if a user attempts to send to an address tied to a
phone number connected with multiple addresses and a different address
was chosen in a previous secure send flow. It always displays the
address chosen for secure send. This updates it so that for any phone
number recipient with multiple addresses, the address currently being
sent to is shown.

### Test plan

Unit tests, manually tested the following scenarios:

1. Send to phone number with multiple addresses (A, B), first time pick
address A on secure send, ensure send confirmation screen displays A
2. Send to address B, ensure send confirmation displays phone number
contact with address B
3. Send to phone number with single address, ensure confirmation screen
displays only contact info and no address
4. Send to address not tied to phone number, ensure confirmation screen
displays only address

### Related issues

N/A

### Backwards compatibility

Yes