refactor(contacts): Use react-native-permissions for contacts (#4979)
### Description

Replace custom contact permissions checking with
react-native-permissions

### Test plan

Unit tests updated

Ran the wallet locally and logged the `contactPermissionStatus` for each
place it was added. Started with contacts permissions turned off and
verified that it logged as something other than `granted`, then turned
permissions to allow and verified that it logged as `granted` in each
place.

### Related issues

- Fixes ACT-1077

### Backwards compatibility

Yes