ci: don't run jobs on main that already run on the merge queue (#5029)
### Description

Now that we have merge queue enabled for main, every PR runs checks at
least twice, once on the PR before being added to the queue and once on
the merge queue. The commit on the merge queue is the same commit that
lands on main, so we don't need to run checks again on main that have
already run on the merge queue (see image below on checks running twice
on commits on main). This disables `check` and `test` workflows on main,
and updates the `e2e-main` workflow to only run tests on OS that have
not already run on the merge queue. This should save us some CI time,
especially on E2E machines.

<img width="527" alt="image"
src="https://github.com/valora-inc/wallet/assets/5062591/68808134-8330-4975-9d85-0626050681d0">


### Test plan

CI

### Related issues

N/A

### Backwards compatibility

N/A

### Network scalability

N/A