fix(keylessBackup): Use existing component for keyless backup delete error (#5045)
### Description

Updates the keyless backup delete error banner to use an existing
component. UI remains unchanged.

### Test plan

Manual and unit tested (existing unit test needed no change).

### Related issues

- Fixes #[issue number here]

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)