test(e2e): increase funded amount for FC tests (#5070)
### Description

FiatConnect e2e tests have been failing with a non enough funds for
transfer out message. This is because the estimated fee seems to be
higher than the balance - amount withdrawn, even though eventually the
actual fee used falls within the balance. So bumping the funded amount
for gas for now to unblock.

We should eventually switch FC to use viem (and get better fee
estimates) and revert this change

### Test plan

Local, CI

### Related issues

- N/A

### Backwards compatibility

N/A