fix: the NFT reward bottom sheet unexpectedly reappears after CTA button is pressed (#5076)
### Description

On the iOS production build the NFT reward bottom sheet unexpectedly
reappears after CTA button is pressed.
The hypothesis is that `onChange` event isn't fired for some reason when
the bottom sheet is closed programmatically.

The fix is to fire `dispatch` on CTA press handler instead of relying on
`onChange` event.

Another possible option is to rely on `onClose` event. But unfortunately
it is called when component unmounts on Android.

### Test plan

Tested manually on:
* Android emulator
* iOS simulator
* dev iOS build on device

Planing to additionally test on the nightly iOS build.

### Related issues

- Related to  RET-1002

### Backwards compatibility

Y

### Network scalability

NA