feat(jumpstart): send jumpstart transactions (#5091)
### Description

This PR adds the plumbing for sending the transactions.
1. add a new redux state variable and actions to track the status of the
transactions. there can only be one in flight at a time. after the
transaction is completed, we should navigate the user to the share
screen and clear the navigation stack.
2. the restriction of one in-flight jumpstart tx at a time comes from
the share link being available only during the send flow. so we should
ensure that the status of the jumpstart transaction is reset when a new
link is generated (this is done in the JumpstartEnterAmount screen)
3. prevent the user from sending a transaction -> navigating back to
change the amount and submitting another transaction, by disabling the
proceed button on the JumpstartEnterAmount screen if there is an
in-flight transaction.
4. add saga to send the transactions. the user should remain on the
confirmation screen until we have a result.
5. add redux migration

Not included in this PR:
1. error handling for when the transactions are not successful (waiting
on design)
2. analytics
6. navigating to the next screen (placeholder for now)

### Test plan


https://github.com/valora-inc/wallet/assets/20150449/376df624-8ea8-49a3-aea5-9a3b2cd0eff2

### Related issues

- Related to RET-994

### Backwards compatibility

Y

### Network scalability

Y