feat(points): Implement Points Home screen (#5128)
### Description

For
[ACT-1022](https://linear.app/valora/issue/RET-1022/[wallet]-build-points-landing-screen-add-statsig-config).
Adds in the Points home screen and sets up Statsig to store points
config information.

See the Statsig Dynamic Config
[here](https://console.statsig.com/4plizaPmWwPL21ASV4QAO0/dynamic_configs/points_config).
Points visibility is gated by the feature gate
[here](https://console.statsig.com/4plizaPmWwPL21ASV4QAO0/gates/show_points).

Note that at the moment, this PR is complete _except_ for:
* The actual Points logo/icon - I'm waiting on a high-quality vector
image before adding it to the branding repo.
* Some copy on the Swap bottom sheet.

I added in the functionality to automatically create "sections" on the
home screen for different point values - we won't need this immediately,
but it was kinda fun to add to figured why not :)

### Test plan

Unit and manual tested. See video below.



https://github.com/valora-inc/wallet/assets/569401/c854b3d3-31c8-4d54-92f2-e65ed6bc2b76


### Related issues

- Fixes
[ACT-1022](https://linear.app/valora/issue/RET-1022/[wallet]-build-points-landing-screen-add-statsig-config).

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [X] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)