feat(jumpstart): Reclaim bottom sheet (#5125)
### Description

This PR adds the reclaim button to the escrow deposits.
It checks if the escrow was claimed and open the bottom sheet, but it
doesn't execute the transaction.

### Test plan

Emulator

### Related issues

- Fixes RET-1003

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [y] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)

---------

Co-authored-by: Kathy Luo <kathyluo18@gmail.com>