feat: Add icon overlay with one Eth asset on Arbitrum or Optimism (#5175)
### Description

Currently, when a user has one balance of Eth, users cannot tell the
difference between on different networks (Ethereum, Arbitrum, Optimism).
This change add an network icon over the token icon for Eth when held on
either Arb or Op.

### Test plan

Tests for presence of TokenIcon and presence / absence of NetworkIcon
depending on network, and passes all tests.

### Related issues

- Fixes ACT-1086

### Backwards compatibility

Passes all previous tests

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will continue to work without code changes but will require
additional testing for new changes