fix: xcode 15.3 build issues by upgrading Flipper version (#5204)
### Description

This PR fixes the iOS build issue with xcode 15.3 by specifying an
explicit Flipper version in the podfile which contains the [fix
needed](https://github.com/facebook/flipper/commit/b3dcdb87f930dbbc9dbacb53ad60996e0111e7d8).
[Context](https://valora-app.slack.com/archives/C025V1D6F3J/p1712220215515279).

Android:
it felt bad to leave the android flipper version at 0.125.0, but 0.250.0
does not work on Android. among other things, the flipper dependencies
which were synced are [no longer synced after
0.182.0](https://stackoverflow.com/a/76222262). so i've just bumped the
version to 0.182.0, which is also what RN 72 uses.

react-native-flipper:
it looks like we are using 2 versions - 0.127 in the dev dependencies,
and 0.212 as a resolution. i'm not sure why this is, we don't need both.
the version used in the dev dependencies is the one that is used to
generate the Pods - when using version 0.212, the iOS release builds
failed locally and in the CI. the failure was something to do with
Flipper
(https://github.com/valora-inc/wallet/actions/runs/8554301604/job/23439350728).
since we were not doing anything with version 0.212 really, i removed
the resolution so that we only have 1 version.

### Test plan

Tested that iOS and Android builds succeed and Flipper works.

### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

Y